Добавление новых проверок, комментариев и прочего

+++ mkiso.sh
@@ -1,6 +1,15 @@
 #!/usr/bin/bash
 # Скрипт для создания LiveUSB образа Calmira GNU/Linux
 # Форк утилиты из Olean Linux
+# (C) 2021 Михаил Краснов <linuxoid85@gmail.com>
+
+## SYNOPSIS
+# mkiso.sh DISTRO OUTPUT
+
+
+##############
+## Checking ##
+##############
 
 if [ $(id -u) != 0 ]; then
 	echo "ERROR: $0 script need to run as root!"
@@ -13,7 +22,7 @@ if [ $# < 2 ]; then
 fi
 
 if [ "$(mksquashfs 2>&1 | grep "Xdict-size")" = "" ]; then
-   echo "ERROR: mksquashfs (from 'squashfs-tools' package) not found or doesn't support -comp xz, aborting, no changes made"
+   echo "ERROR: mksquashfs (from 'squashfs-tools' package) not found or doesn't support xz compressing mode, aborting, no changes made"
    echo "You may consider installing squashfs-tools package"
    exit 1
 fi
@@ -32,7 +41,12 @@ ISOLINUXDIR=$CWD/livecd/isolinux
 DISTRONAME="Calmira LX4"
 LABEL=CalmiraLiveCD
 CALM_ROOT=$1
-#OUTPUT="calmira-$CALMVERSION.iso"
+
+if [ -z "$2" ]; then
+	OUTPUT="calmira-$CALMVERSION.iso"
+else
+	OUTPUT="$2"
+fi
 
 isolinux_files="chain.c32 isolinux.bin ldlinux.c32 libutil.c32 reboot.c32 menu.c32
 isohdpfx.bin isolinux.cfg libcom32.c32 poweroff.c32"
@@ -54,6 +68,7 @@ unset file
 rm    $WDIR
 mkdir $WDIR
 
+
 # prepare isolinux in working dir
 mkdir $WDIR/{filesystem,isolinux,boot}
 
@@ -66,14 +81,22 @@ echo "$DISTRONAME" > $WDIR/isolinux/venomlive
 
 cp $CALM_ROOT/boot/{vmlinuz-*-calm-kernel,vmlinuz} $WDIR/boot/
 
-mksquashfs $CALM_ROOT root.sfs                             \
-	-b 1048576 -comp xz -Xdict-size 100%        \
-	-e $CALM_ROOT/tmp/*                         \
+# Создание и хранение root.sfs (сквоша системы) сделано в текущей
+# директории на всякий случай, чтобы при изменении в файлах
+# загрузчика (и прочих не относящихся к системе) не пересобирать
+# всю систему в сквош, а скопировать готовый
+mksquashfs $CALM_ROOT root.sfs           \
+	-b 1048576 -comp xz -Xdict-size 100% \
+	-e $CALM_ROOT/tmp/*                  \
 	-e $CALM_ROOT/usr/src/*
 	
 cp root.sfs $WDIR/filesystem
 
-rm -rf $OUTPUT
+if [ -f "$OUTPUT" ]; then
+	rm -rf $OUTPUT
+fi
+
+# Make iso
 xorriso -as mkisofs                              \
 		-r -J -joliet-long                       \
 		-l -cache-inodes                         \
